ensure poetry authors are parsed properly (fixes #2665)
