+++ news/2665.bugfix.md
@@ -1,0 +1,2 @@
+Fix errors when parsing poetry format that contains special characters in author name.
+Poetry-specific `parse_name_email` and `NAME_EMAIL_RE` moved from `pdm.formats.base` to `pdm.formats.poetry`.
